[Transform] refactor transform functions that utilize composite aggs (#67986)
Many of the queries and APIs with the composite agg functions are very similar (if not exactly the same).

This commit refactors both the pivot and latest transform functions so that they are more uniform in design and common functions are extracted out.
