+++ x-pack/plugin/transform/src/test/java/org/elasticsearch/xpack/transform/transforms/common/DocumentConversionUtilsTests.java
@@ -10,6 +10,7 @@ import org.elasticsearch.action.fieldcaps.FieldCapabilities;
 import org.elasticsearch.action.fieldcaps.FieldCapabilitiesResponse;
 import org.elasticsearch.action.index.IndexRequest;
 import org.elasticsearch.common.Strings;
+import org.elasticsearch.common.collect.MapBuilder;
 import org.elasticsearch.test.ESTestCase;
 
 import java.util.Collections;
@@ -25,55 +26,65 @@ import static org.hamcrest.Matchers.oneOf;
 
 public class DocumentConversionUtilsTests extends ESTestCase {
 
-    private static String INDEX = "some-index";
-    private static String PIPELINE = "some-pipeline";
-    private static String ID = "some-id";
-    private static Map<String, Object> DOCUMENT = Collections.unmodifiableMap(
-        new HashMap<String, Object>() {{
-            put("_id", ID);
-            put("field-1", "field-1-value");
-            put("field-2", "field-2-value");
-            put("field-3", "field-3-value");
-            put("_internal-field-1", "internal-field-1-value");
-            put("_internal-field-2", "internal-field-2-value");
-        }});
-    private static Map<String, Object> DOCUMENT_WITHOUT_INTERNAL_FIELDS = Collections.unmodifiableMap(
-        new HashMap<String, Object>() {{
-            put("field-1", "field-1-value");
-            put("field-2", "field-2-value");
-            put("field-3", "field-3-value");
-        }});
+    private static final String INDEX = "some-index";
+    private static final String PIPELINE = "some-pipeline";
+    private static final String ID = "some-id";
+    private static final Map<String, Object> DOCUMENT = Collections.unmodifiableMap(
+        MapBuilder.<String, Object>newMapBuilder()
+            .put("field-1", "field-1-value")
+            .put("field-2", "field-2-value")
+            .put("field-3", "field-3-value")
+            .put("_internal-field-1", "internal-field-1-value")
+            .put("_internal-field-2", "internal-field-2-value")
+            .map()
+    );
+    private static final Map<String, Object> DOCUMENT_WITHOUT_ID = Collections.unmodifiableMap(
+        MapBuilder.<String, Object>newMapBuilder()
+            .put("field-1", "field-1-value")
+            .put("field-2", "field-2-value")
+            .put("field-3", "field-3-value")
+            .put("_internal-field-1", "internal-field-1-value")
+            .put("_internal-field-2", "internal-field-2-value")
+            .map()
+    );
+    private static final Map<String, Object> DOCUMENT_WITHOUT_INTERNAL_FIELDS = Collections.unmodifiableMap(
+        MapBuilder.<String, Object>newMapBuilder()
+            .put("field-1", "field-1-value")
+            .put("field-2", "field-2-value")
+            .put("field-3", "field-3-value")
+            .map()
+    );
 
     public void testConvertDocumentToIndexRequest_MissingId() {
         Exception e =
             expectThrows(
                 Exception.class,
-                () -> DocumentConversionUtils.convertDocumentToIndexRequest(Collections.emptyMap(), INDEX, PIPELINE));
+                () -> DocumentConversionUtils.convertDocumentToIndexRequest(null, Collections.emptyMap(), INDEX, PIPELINE));
         assertThat(e.getMessage(), is(equalTo("Expected a document id but got null.")));
     }
 
     public void testConvertDocumentToIndexRequest() {
-        IndexRequest indexRequest = DocumentConversionUtils.convertDocumentToIndexRequest(DOCUMENT, INDEX, PIPELINE);
+        IndexRequest indexRequest = DocumentConversionUtils.convertDocumentToIndexRequest(ID, DOCUMENT, INDEX, PIPELINE);
         assertThat(indexRequest.index(), is(equalTo(INDEX)));
         assertThat(indexRequest.id(), is(equalTo(ID)));
         assertThat(indexRequest.getPipeline(), is(equalTo(PIPELINE)));
-        assertThat(indexRequest.sourceAsMap(), is(equalTo(DOCUMENT_WITHOUT_INTERNAL_FIELDS)));
+        assertThat(indexRequest.sourceAsMap(), is(equalTo(DOCUMENT_WITHOUT_ID)));
     }
 
     public void testConvertDocumentToIndexRequest_WithNullIndex() {
-        IndexRequest indexRequest = DocumentConversionUtils.convertDocumentToIndexRequest(DOCUMENT, null, PIPELINE);
+        IndexRequest indexRequest = DocumentConversionUtils.convertDocumentToIndexRequest(ID, DOCUMENT, null, PIPELINE);
         assertThat(indexRequest.index(), is(nullValue()));
         assertThat(indexRequest.id(), is(equalTo(ID)));
         assertThat(indexRequest.getPipeline(), is(equalTo(PIPELINE)));
-        assertThat(indexRequest.sourceAsMap(), is(equalTo(DOCUMENT_WITHOUT_INTERNAL_FIELDS)));
+        assertThat(indexRequest.sourceAsMap(), is(equalTo(DOCUMENT_WITHOUT_ID)));
     }
 
     public void testConvertDocumentToIndexRequest_WithNullPipeline() {
-        IndexRequest indexRequest = DocumentConversionUtils.convertDocumentToIndexRequest(DOCUMENT, INDEX, null);
+        IndexRequest indexRequest = DocumentConversionUtils.convertDocumentToIndexRequest(ID, DOCUMENT, INDEX, null);
         assertThat(indexRequest.index(), is(equalTo(INDEX)));
         assertThat(indexRequest.id(), is(equalTo(ID)));
         assertThat(indexRequest.getPipeline(), is(nullValue()));
-        assertThat(indexRequest.sourceAsMap(), is(equalTo(DOCUMENT_WITHOUT_INTERNAL_FIELDS)));
+        assertThat(indexRequest.sourceAsMap(), is(equalTo(DOCUMENT_WITHOUT_ID)));
     }
 
     public void testRemoveInternalFields() {
